S-00006 - __pycache__ build cleanUP.py dist howToExcute.txt main.kv main.py main.spec replayScript.json Scripts venv Fixed SSubCategoryGroups. __pycache__ build cleanUP.py dist howToExcute.txt main.kv main.py main.spec replayScript.json Scripts venv Fixed sEditorVisible. __pycache__ build cleanUP.py dist howToExcute.txt main.kv main.py main.spec replayScript.json Scripts venv Added API request section.

+++ Scripts/addingUsers/demoData.py
@@ -27,10 +27,13 @@ def copy_users(relevant_user, core_path, solution_qa_path):
         pass
     
     for user in relevant_user:
-        try:
-            shutil.copytree(core_demo_data_path + "\\" + user, solution_demodata + "\\" + user)
-        except:
-            raise Exception("insight use case is not presented in the core: ", user)
+        if os.path.exists(solution_demodata + "\\" + user):
+            print("User: " + user + " is all ready in the solution level")
+        else:
+            try:
+                shutil.copytree(core_demo_data_path + "\\" + user, solution_demodata + "\\" + user)
+            except:
+                raise Exception("insight use case is not presented in the core: " + user)
 
 
 def modify_user(solution_qa_path, user, local_currency, foreign_currency, country_name, country_code):
@@ -55,20 +58,30 @@ def modify_user(solution_qa_path, user, local_currency, foreign_currency, countr
             return current_country
     
     def update_column(column, file, func):
-        # df = readCsv(solution_qa_path + "\\" + user + "\\" + file)
+        reWrite = False
         try:
             df = readCsv(solution_qa_path + "\\" + user + "\\" + file)
             for i in range(len(df[column])):
+                check = df.loc[i, column]
                 df.loc[i, column] = func(df[column][i])
-            writeCsv(solution_qa_path + "\\" + user + "\\" + file, df)
+                if check == df.loc[i, column]:
+                    reWrite = True
+            if reWrite:
+                writeCsv(solution_qa_path + "\\" + user + "\\" + file, df)
         except:
             print(solution_qa_path + "\\" + user + "\\" + file, "col " + column + " does not exist")
     
-    update_column("currencyCd", "DAccount.csv", update_currency)
-    update_column("currencyCd", "DTransaction.csv", update_currency)
-    update_column("currencyCdOriginal", "DTransaction.csv", update_currency)
-    update_column("countryCd", "DLocation.csv", update_country_code)
-    update_column("countryName", "DLocation.csv", update_country_name)
+    for thisFile in os.listdir(solution_qa_path + "\\" + user):
+        csvFile = readCsv(solution_qa_path + "\\" + user + "\\" + thisFile)
+        for col in ['currencyCd', 'currencyCdOriginal', 'countryCd', 'countryName']:
+            if col in csvFile:
+                if 'currency' in col:
+                    update_column(col, thisFile, update_currency)
+                else:
+                    if 'Cd' in col:
+                        update_column(col, thisFile, update_country_code)
+                    elif 'Name' in col:
+                        update_column(col, thisFile, update_country_name)
 
 
 def modify_users_in_solution(solution_qa_path, users, input_json):
@@ -84,7 +97,8 @@ def main(argv):
     extraUsers = getCol(argv[5], 'USERS')
     relevant_user = find_relevant_users(core, extraUsers, argv[3], argv[4])
     copy_users(relevant_user, core, product)
-    modify_users_in_solution(product, os.listdir(product + "\DemoData"), argv[2])
+    print("Copying users finished")
+    modify_users_in_solution(product, relevant_user, argv[2])
     print("Overwriting finished")
 
 
@@ -97,3 +111,5 @@ if __name__ == "__main__":
 # 3 B_users checkBox bool
 # 4 modified checkBox bool
 # 5 file name
+
+# os.listdir(product + "\\DemoData")
